system_tray: move into the items module
SystemTray is a native item like Empty, Flickable, Path, etc., so house
it alongside the other item modules under items/ instead of carrying a
top-level `system_tray` namespace in the crate root. No behavioural
change.
