+++ internal/core/items/system_tray.rs
