+++ internal/core/items/system_tray/ksni.rs
