+++ internal/core/items/system_tray/tray_icon.rs
